merge preprod into Social monitor

+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/services/settingsmanagement/SettingsManager.java
@@ -1,7 +1,9 @@
 package com.realtech.socialsurvey.core.services.settingsmanagement;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
+
 import com.realtech.socialsurvey.core.entities.SettingsDetails;
 import com.realtech.socialsurvey.core.enums.OrganizationUnit;
 import com.realtech.socialsurvey.core.enums.SettingsForApplication;
@@ -12,7 +14,7 @@ import com.realtech.socialsurvey.core.services.settingsmanagement.impl.InvalidSe
  */
 public interface SettingsManager {
 
-	public Map<String, Double> calculateSettingsScore(List<SettingsDetails> settingsDetails);
+	public Map<String, BigInteger> calculateSettingsScore(List<SettingsDetails> settingsDetails);
 
 	public List<SettingsDetails> getScoreForCompleteHeirarchy(long companyId, long branchId, long regionId);
